seperated firebase init.js
